requirements.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.8.0
click==8.1.7
colorama==0.4.6
coverage==7.6.1
decorator==5.1.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
inflection==0.5.1
iniconfig==2.0.0
ipython==8.26.0
isort==5.13.2
jedi==0.19.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
platformdirs==4.3.2
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
setuptools==73.0.1
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.1
uritemplate==4.1.1
wcwidth==0.2.13